+++ pkg/balancer/node/types.go
@@ -34,7 +34,7 @@ type NodeMetrics struct {
 	Lock           sync.Mutex `json:"-"`
 	Health         string     `json:"health"`
 	ResponseTime   float32    `json:"response_time"`
-	Connections    int        `json:"connections"`
+	Connections    uint32     `json:"connections"`
 	CreatedNewNode bool       `json:"-"`
 	// LastRequestTime time.Time  `json:"last_request"`
 }
@@ -45,5 +45,5 @@ type NodeQueue struct {
 	connChan      chan *types.Connection // Signal channel for new connections
 	closeSignal   chan struct{}          // Signal channel for closing the queue
 	workChan      chan *types.Connection // connections to send to the worker pool
-	workerThreads uint
+	workerThreads uint16
 }
